requirements.txt
boto3
python-nmap
sslyze
scapy
whois
sublist3r
dnspython==1.15.0
py2neo
psutil
mysql-connector-python
requests_toolbelt
beautifulsoup4
google
censys
shodan
fire
pyshark
progressbar
fpdf
